+++ io.github.adr.embedded/src/main/java/io/github/adr/embedded/YStatementJustifications.java
@@ -28,7 +28,7 @@ import java.lang.annotation.Target;
     ElementType.TYPE,
     ElementType.TYPE_PARAMETER,
     ElementType.TYPE_USE})
-@Retention(RetentionPolicy.RUNTIME)
+@Retention(RetentionPolicy.SOURCE)
 @Documented
 public @interface YStatementJustifications {