+++ CHANGELOG.md
@@ -9,12 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Support for Java9
 - Refined example in `README.md`
+
+### Changed
+
+- Updated to Java 21
+- Change to `RetentionPolicy.SOURCE`
 
 ### Fixed
 
-- Fix typo in `ArchitecturallySignificant` annotation
+- Fix typo in `ArchitecturallySignificant` and `ArchitecturalStyle`
 
 ## [1.0.0] – 2017-09-10